Fixup timestamp representation in sdk encoding
It's a semantic no-op (except in the case where we provide an invalid
value to the sdk encoder that's greater than can be represented in
uint40). Mostly did this to make sure the code is consistent.

+++ sdk/market.js
@@ -117,7 +117,7 @@ function bidStructHash(bid) {
       [
         "bytes32",
         "uint256",
-        "uint256",
+        "uint40",
         "address",
         "uint256",
         "address",
@@ -158,7 +158,7 @@ function askStructHash(ask) {
       [
         "bytes32",
         "uint256",
-        "uint256",
+        "uint40",
         "address",
         "uint256",
         "address",
